requirements.txt
coverage==7.5.3
Django==5.0.6
django-cors-headers==4.3.1
django-filter==24.2
djangorestframework==3.15.1
python-dotenv==1.0.1
psycopg2==2.9.9
